RPA for the supervision of motorways to identify traffic violators with insurance/registration backlogs, stolen vehicles, fine defaulters and also notifying major accidents to the nearby hospital, owner and police patrol.
It is well-known that non-compliance of traffic rules and other violations has been increasing constantly during the recent years. Though traffic control devices and traffic guards have control within city limits, in highways it is a significant problem. Hence the proposed prototype automates supervision of motorways for traffic violation and accidents. A dummy website mimicking traffic information is hosted using WebHost000 is designed for insurance details and police records. The automation is designed as a Snippet workflow in UiPath Studio.